+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 - Drop support for GHCs prior 8.6
 - Require libpq >=10.22 (when using pkgconfig).
+- Use CApiFFI for some parts of the API
+  (we cannot use for all due https://gitlab.haskell.org/ghc/ghc/-/issues/22043)
 
 0.9.4.3
 -------